% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceData.R
\name{BalanceData}
\alias{BalanceData}
\title{Clean text and buil term matrix for bag of words model or TF DFI.}
\usage{
BalanceData(dataset)
}
\arguments{
\item{x}{unbalanced dataset, a dataframe : two column: first reviews and second binary class label: negative =0 and positive=1.}
}
\value{
balanced dataset, save as out.tsv name in inst folder
}
\description{
Clean text and buil term matrix for bag of words model or TF DFI.
}
\examples{
library(SentiAnalyzer)
imbalance_data<- read.delim(system.file(package = "SentiAnalyzer", "extdata/Imbalance_Restaurant_Reviews.tsv"),quote='',stringsAsFactors = FALSE)
BalanceData(imbalance_data)
}
\author{
Zahra Khoshmanesh
}
