% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPrediction.R
\name{BuildPrediction}
\alias{BuildPrediction}
\title{predict the classes using trained algorithms and give confusion matrix  in return}
\usage{
BuildPrediction(x)
}
\arguments{
\item{x}{input is a dataframe for document-term matrix in which columns are the terms and row are binary variable 1 if that term exist in that data instance}
}
\value{
list of 4 confusion matrix for each trained classification algorithm
}
\description{
predict the classes using trained algorithms and give confusion matrix  in return
}
\examples{
library(SentiAnalyzer)
csv_data <- read.csv(system.file(package = "SentiAnalyzer", "extdata/testing.csv"))
my_training_data <- BuildPrediction(csv_data)
}
\author{
Atousa Zarindast
}
