% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_dataset}
\alias{clean_dataset}
\title{clean_dataset}
\format{A data frame with 1000 rows and 97 variables:
\describe{
  \item{X}{row number}
  \item{also}{...}
  ...
}}
\usage{
clean_dataset
}
\description{
A dataset containing ...
}
\keyword{datasets}
