% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanText.R
\name{CleanText}
\alias{CleanText}
\title{Clean text and buil term matrix for bag of words model.}
\usage{
CleanText()
}
\arguments{
\item{x}{A tsv file having two columns, review as text, label as binary.}
}
\value{
The term matrix converted to dataframe.
}
\description{
Clean text and buil term matrix for bag of words model.
}
