% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{compare the performance of the 4 algorithms}
\usage{
comparison(x)
}
\arguments{
\item{x}{input is a dataframe for document-term matrix in which columns are the terms and row are binary variable 1 if that term exist in that data instance}
}
\value{
visualization of trained algorithms (for classifiying textual data), comparision and select the best!
}
\description{
compare the performance of the 4 algorithms
}
\examples{
library(SentiAnalyzer)
csv_data <- read.csv(system.file(package = "SentiAnalyzer", "extdata/testing.csv"))
my_training_data <- BuildTraining(csv_data)
}
\author{
Atousa Zarindast
}
