% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{original_dataset}
\alias{original_dataset}
\title{original_dataset}
\format{A data frame with 1000 rows and 2 variables:
\describe{
  \item{X}{row number}
  \item{also}{...}
  ...
}}
\usage{
original_dataset
}
\description{
A dataset containing ...
}
\keyword{datasets}
