% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orignal_dataset}
\alias{orignal_dataset}
\title{orignal_dataset}
\format{A data frame with 1000 rows and 2 variables:
\describe{
  \item{Liked}{class variable}
  \item{Review}{text reviews}
  ...
}}
\usage{
orignal_dataset
}
\description{
A dataset containing ...
}
\keyword{datasets}
