% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing}
\alias{testing}
\title{Testing}
\format{A data frame with 1000 rows and 98 variables:
\describe{
  ...
}}
\usage{
testing
}
\description{
A reduced restaurant review dataset containing 98 terms of 1000 observation
}
\keyword{datasets}
