% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing1}
\alias{testing1}
\title{testing1}
\format{A data frame with 100 rows and 4 variables:
\describe{
  ...
}}
\usage{
testing1
}
\description{
A dataset containing ...
}
\keyword{datasets}
