% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_models}
\alias{trained_models}
\title{trained_models}
\format{A list with  5 trained algorithms resampling summaries: accuracy, kappa
\describe{ list, in sequence: GBM, KNN, NB, RF,SVM_Poly 
  ...
}}
\usage{
trained_models
}
\description{
output of BuildTraining, lists
}
\keyword{datasets}
