% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{VisualizeData}
\alias{VisualizeData}
\title{Visualize dataset and get some insight of data.}
\usage{
VisualizeData(dataset, termcount)
}
\arguments{
\item{dataset}{dataset.}

\item{termcount}{termcount:used for filtering the highest terms repeated in reviews usually > 10}
}
\value{
some diagram and insight of data

A ggplot
}
\description{
Visualize dataset and get some insight of data.
}
\examples{
\dontrun{
library(SentiAnalyzer)
direction <- system.file(package = "SentiAnalyzer", "extdata/Restaurant_Reviews.tsv")
original_dataset <- read.delim(direction,quote='',stringsAsFactors = FALSE)
VisualizeData(dataset=original_dataset,termcount=15)}
}
\author{
Zahra Khoshmanesh
}
