% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceData.R
\name{BalanceData}
\alias{BalanceData}
\title{Clean text and buil term matrix for bag of words model or TF DFI.}
\usage{
BalanceData(x)
}
\arguments{
\item{x}{unbalanced dataset.}
}
\value{
balanced dataframe
}
\description{
Clean text and buil term matrix for bag of words model or TF DFI.
}
\author{
Zahra Khoshmanesh
}
