% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceData.R
\name{BalanceData}
\alias{BalanceData}
\title{Clean text and buil term matrix for bag of words model or TF DFI.}
\usage{
BalanceData(dataset)
}
\arguments{
\item{x}{unbalanced dataset, a tsv format : tab delimeter, two column: first text and second binary class label.}
}
\value{
balanced dataset, save as out.tsv name in inst folder
}
\description{
Clean text and buil term matrix for bag of words model or TF DFI.
}
\examples{
BalanceData('./inst/Imbalance_Restaurant_Reviews.tsv')
}
\author{
Zahra Khoshmanesh
}
