% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPrediction.R, R/BuildPrediction_jb.R
\name{BuildPrediction}
\alias{BuildPrediction}
\title{predict the classes using trained algorithms and give cnfusion matrix  in return}
\usage{
BuildPrediction(list)

BuildPrediction(list)
}
\arguments{
\item{x}{input is a dataframe for bag of word file in which columns are the terms and row are binary variable 1 if that term exist in that data instance}

\item{x}{input is a dataframe for bag of word file in which columns are the terms and row are binary variable 1 if that term exist in that data instance}
}
\value{
3 confusion matrix for each trained classification algorithm

3 confusion matrix for each trained classification algorithm
}
\description{
predict the classes using trained algorithms and give cnfusion matrix  in return

predict the classes using trained algorithms and give cnfusion matrix  in return
}
\author{
Atousa Zarindast

Atousa Zarindast
}
