% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTraining.R
\name{BuildTraining}
\alias{BuildTraining}
\title{Train Machine learning algorithms and give trained models}
\usage{
BuildTraining(x)
}
\arguments{
\item{x}{input is a dataframe for bag of word file in which columns are the terms and row are binary variable 1 if that term exist in that data instance}
}
\value{
traned models
}
\description{
Train Machine learning algorithms and give trained models
}
\author{
Atousa Zarindast
}
