% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_dataset}
\alias{clean_dataset}
\title{clean_dataset}
\format{A data frame with 1000 rows and 97 variables:
\describe{ cleaned dataset term matrix
  ...
}}
\usage{
clean_dataset
}
\description{
output of CleanText, the corpus that is cleaned up and stemmed for base words
}
\keyword{datasets}
