% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanText.R
\name{CleanText}
\alias{CleanText}
\title{Clean text and buil term matrix for bag of words model or TF DFI.}
\usage{
CleanText(source_dataset, dtm_method)
}
\arguments{
\item{source_dataset}{A tsv file having two columns, review as text, label as binary.}

\item{dtm-method}{1 for bag of word, 2 for TF DFI.}
}
\value{
The term matrix converted to dataframe.
}
\description{
Clean text and buil term matrix for bag of words model or TF DFI.
}
