% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanText.R
\name{CleanText}
\alias{CleanText}
\title{Clean text and build term matrix for bag of words,TF DFI and bi-gram.}
\usage{
CleanText(source_dataset, dtm_method, reductionrate)
}
\arguments{
\item{source_dataset}{A dataframe  having two columns, review as text, label as binary.}

\item{dtm_method}{1 for bag of word, 2 for TF DFI, 3 for bigram.}

\item{reductionrate}{how many percent of term matrix you want to keep,usually 0.999 and not less than 0.99.}
}
\value{
dataframe "dataset" : The term matrix converted to dataframe plus target label.
}
\description{
Clean text and build term matrix for bag of words,TF DFI and bi-gram.
}
\examples{
\dontrun{
library("SentiAnalyzer")
direction <- system.file(package = "SentiAnalyzer", "extdata/Restaurant_Reviews.tsv")
orignal_dataset <- read.delim(direction,quote='',stringsAsFactors = FALSE)
CleanText(original_dataset,dtm_method=1,reductionrate=0.99)
CleanText(original_dataset,dtm_method=2,reductionrate=0.99)
CleanText(original_dataset,dtm_method=3,reductionrate=0.999)}
}
\author{
Zahra Khoshmanesh
}
