% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{visualize performance of the algorithm}
\usage{
comparison(list)
}
\arguments{
\item{x}{input is a dataframe for bag of word file in which columns are the terms and row are binary variable 1 if that term exist in that data instance}
}
\value{
visualization of trained algrithms (for classifiying textual data), comparision and select the best!
}
\description{
visualize performance of the algorithm
}
\author{
Atousa Zarindast
}
