% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imbalance_data}
\alias{imbalance_data}
\title{imbalance_data}
\format{A data frame with 903 rows and 2 variables:
\describe{
  \item{Liked}{class variable}
  \item{Review}{text reviews}
  ...
}}
\usage{
imbalance_data
}
\description{
A dataset containing ...
}
\keyword{datasets}
