% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{original_dataset}
\alias{original_dataset}
\title{original_dataset}
\format{A data frame with 1000 rows and 2 variables:
\describe{
  \item{Liked}{class variable}
  \item{Review}{text reviews}
  ...
}}
\source{
\url{https://www.kaggle.com/c/restaurant-reviews}
}
\usage{
original_dataset
}
\description{
A dataset containing restaurant review data
}
\keyword{datasets}
