% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing1}
\alias{testing1}
\title{Testing1}
\format{Reduced dataframe for testing purpose with 1000 rows and 8 variables:
\describe{
  ...
}}
\usage{
testing1
}
\description{
A further reduced restaurant review dataset containing 8 terms of 1000 observation
}
\keyword{datasets}
