% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing1}
\alias{testing1}
\title{Testing1}
\format{A data frame with 1000 rows and 98 variables:
\describe{
  \item{X}{row number}
  \item{also}{...}
  ...
}}
\usage{
testing1
}
\description{
A dataset containing ...
}
\keyword{datasets}
