% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_models}
\alias{trained_models}
\title{trained_models}
\format{A data frame:
\describe{ trained ML models in as list variables
  ...
}}
\usage{
trained_models
}
\description{
A dataset containing ...
}
\keyword{datasets}
