% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{VisualizeData}
\alias{VisualizeData}
\title{Visualize dataset and get some insight of data.}
\usage{
VisualizeData(dataset, termcount)
}
\arguments{
\item{x}{dataset.}

\item{y}{termcount:used for filtering the highest terms repeated in reviews usually > 10}
}
\value{
some diagram and insight of data
}
\description{
Visualize dataset and get some insight of data.
}
\examples{
library(SentiAnalyzer)
direction <- system.file(package = "SentiAnalyzer", "extdata/Restaurant_Reviews.tsv")
orignal_dataset <- read.delim(direction,quote='',stringsAsFactors = FALSE)
VisualizeData(orignal_dataset,15)
}
\author{
Zahra Khoshmanesh
}
