% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeData.R
\name{VisualizeData}
\alias{VisualizeData}
\title{Visualize dataset and get some insight of data.}
\usage{
VisualizeData(dataset)
}
\arguments{
\item{x}{dataset.}
}
\value{
some diagram and insight of data
}
\description{
Visualize dataset and get some insight of data.
}
\examples{
VisualizeData('./inst/Restaurant_Reviews.tsv')
}
\author{
Zahra Khoshmanesh
}
